+++ lute/cli/include/lute/toolchain.h
@@ -2,7 +2,5 @@
 
 #include "lute/reporter.h"
 
-#include <string>
-
-// Return true on successful dispatch.
-bool dispatchToPinnedLute(int argc, char** argv, const std::string& currentExe, LuteReporter& reporter);
+// Re-execs into the project-pinned lute when one is configured. Returns false on failure.
+bool dispatchToPinnedLute(int argc, char** argv, LuteReporter& reporter);
